make files lowercase
