+++ .gitignore
@@ -9,3 +9,6 @@ setup.h
 .vscode/c_cpp_properties.json
 .vscode/launch.json
 src/setup.h
+src/setup_.h
+
+ls | while read upName; do loName=`echo "${upName}" | tr '[:upper:]' '[:lower:]'`; git mv "$upName" "$loName"; done
