+++ lib/libLightningUtils/transitioneffect.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <stdint.h>
-#include "Effect.h"
+#include "effect.h"
 #include <hsb.h>
 
 /**
